Separed Classes BookingManager and the Interface
