+++ TestNinja/TestNinja/Mocking/IBookingManager.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections.Generic;
+
+namespace TestNinja.Mocking
+{
+    public interface IBookingManager
+    {
+        IEnumerable<Booking> GetActiveBookingList(Booking booking);
+    }
+}
